+++ student.rb
@@ -1,6 +1,6 @@
-# require_relative './person'
+require_relative './person'
 class Student < Person
-  def initialize(classroom)
+  def initialize(classroom, name, age)
     super(name, age)
     @classroom = classroom
   end
